refactor(pr): use dynamic import for subcommand
