+++ packages/opencode/src/index.ts
@@ -27,7 +27,7 @@ import { TuiThreadCommand } from "./cli/cmd/tui/thread"
 import { AcpCommand } from "./cli/cmd/acp/command"
 import { EOL } from "os"
 import { WebCommand } from "./cli/cmd/web"
-import { PrCommand } from "./cli/cmd/pr"
+import { PrCommand } from "./cli/cmd/pr/command"
 import { SessionCommand } from "./cli/cmd/session"
 import { DbCommand } from "./cli/cmd/db/command"
 import path from "path"
